% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eofinder.R
\name{collect_responses}
\alias{collect_responses}
\title{Collect the response of a query to eofinder}
\usage{
collect_responses(query)
}
\arguments{
\item{query}{A charachter with the API query parameters}
}
\value{
An sf object
}
\description{
Internal helper thus not exported
}
\keyword{internal}
