% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eofinder.R
\name{download_S2_codede}
\alias{download_S2_codede}
\title{Download matched acquisitions from CODE-DE}
\usage{
download_S2_codede(s2_matches = NULL, key = NULL, outdir = ".")
}
\arguments{
\item{s2_matches}{An sf object returned from calling \code{query_s2_codede()}.}

\item{key}{A character vector indicating an CODE-DE API key.}

\item{outdir}{A character vector indicating the directory where the downloads
should be written to.}
}
\value{
Nothing, called for its side effect.
}
\description{
Use this function to download matching Sentinel-2 acquisitions from
the CODE-DE data storage to your local machine. This functions assumes that
specify an API key which requires you to have a subscription with CODE-DE.
Currently, the easiest way to obtain such an API key is to head over to
\url{https://finder.code-de.org/}, log-in with your user account, query some
spatio-temporal extent and copy the key from the URL results.
}
