% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eofinder.R
\name{parse_item}
\alias{parse_item}
\title{Convert eofinder item response to sf}
\usage{
parse_item(item)
}
\arguments{
\item{item}{A list object represening a single item response from eofinder}
}
\value{
An sf object
}
\description{
Internal helper thus not exported
}
\keyword{internal}
