% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eofinder.R
\name{query_s2_codede}
\alias{query_s2_codede}
\title{Query Sentinel-2 scenes from eofinder}
\usage{
query_s2_codede(
  aoi = NULL,
  level = NULL,
  product = NULL,
  cloudcover = 100,
  start = "2015-01-01 00:00:00 CET",
  end = Sys.time()
)
}
\arguments{
\item{aoi}{An unprojected sf object. Its bounding box will be used to
query th eofinder API.}

\item{level}{The processing level of the collection to query. Available
values are "LEVEL1", "LEVEL2A", "LEVEL2AP", and "LEVEL3".}

\item{product}{The product to query. Available options are "L1C", "L2A",
"L2A-MAJA", "L2A-FORCE", and "L3-WASP", depending on the selected level.}

\item{cloudcover}{A numeric giving the maximum allowed cloud cover.}

\item{start}{A character or date object indicating the start of the
temporal window for the query.}

\item{end}{A character or date object indicating the end of the
temporal window for the query.}
}
\value{
An sf object with one matching acquisition per row.
}
\description{
Using the eofinder API, this function can be used to query
matching Sentinel-2 scenes with a user-defined spatio-temporal
extent. This functions returns an sf object with additional
meta data for each acqusition which can be used for further filtering
of the results.
}
